Update Submit-CertificateRequest.ps1
Adds the functionality to accept text contents of a certificate signing request. This is particularly useful when you would like to use this Cmdlet somewhere that you do not have access to a file path or do not want to create a temporary file (such as on a MID server). I implemented this to complete a project requirement but thought it may be useful to others. Thank you for this module, it is very useful.